pyproject.toml
[project]
name = "learning-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "psycopg[binary,pool]>=3.2.10",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
]
